197: calculate avarage user stats
