+++ prisma/schema.prisma
@@ -30,6 +30,9 @@ model User {
   results       Result[]
   achievements  UserAchievement[]
 
+  avarageAccuracy Decimal @default(0) @db.Decimal(5, 2)
+  avarageCpm      Decimal @default(0) @db.Decimal(6, 2)
+
   role UserRole @default(USER)
 
   snippets     Snippet[]
